+++ CHANGELOG.md
@@ -30,7 +30,10 @@ This project adheres to [Semantic Versioning](http://semver.org/).
   - "sec. iv - 1/2 sec. ii p. chr"
   - "între mil. iii şi 1/2 mil. ii p. chr"
 - Added a safety guard to the TimeExpression constructor
-- Prevented matching of day-month expressions like "23 decembrie", "30 august", etc. and hour-minute expressions like "18.05"
+- Prevented matching the following patterns:
+  - day-month expressions like "23 decembrie", "30 august", etc.
+  - hour-minute expressions like "18.05"
+  - term-value like "luna 10", "minutul 30", etc.
 
 ## 1.7
 - Added support for preserving diacritics of transformed values